Barebones First Commit
Initial file dump, not even close to a version 1.0 here. These scripts
represent the three main parts of the script.

renderfarm.py- the input side of things- designed to be run in maya,
will eventually be capable of  reading info directly from the shotinfo
document from px

jobstack- the brains of the system - designed to query networked
machines and distribute jobs

serverscript - this script contains the subrocess command used to do the
actual rendering, (thats about the only part of this that's working)
this script needs major retooling to match up with the protocols im
using in the two previous scripts